Delete and recreate netStats directory properly
Potential fix for negative network stats. It looks like the intent of
resetStats() is to remove the netStats directory and recreate it,
however the call to delete() will not delete the directory if it is not
empty (and it won't be if statistics have been previously written to
disk).

Implement a "remove entire directory tree" method and use it instead of
delete().

NOTE: This commit has not currently been tested (still resolving some
unrelated build issues)